fix: resolve all compilation errors and improve type safety
- Fixed type conversions for AlertSeverity and AlertStatus
- Stubbed out missing fields in enhanced_keyword_classifier and monitoring_integration_service
- Fixed risk service type mismatches (RiskFactorDetail to RiskScore conversions)
- Added convertRiskRecommendationsToDetails helper function
- Fixed CalculateFactor to CalculateEnhancedFactor method calls
- Fixed trend analysis to use AnalyzeTrends with proper request structure
- Fixed field access issues in unified_audit_service and enhanced_risk_service
- Removed unused imports
- Fixed UserFeedback data mapping in feedback service
- Improved error handling and type safety across codebase

All compilation errors resolved. Server builds successfully.

+++ internal/feedback/developer_feedback_collector.go
@@ -293,7 +293,7 @@ func (dfc *DeveloperFeedbackCollector) calculateROI(feedback *DeveloperFeedback)
 		debtAdjustment = 0.4
 	}
 
-	adjustedScore := score * debtAdjustment
+	adjustedScore := float64(score) * debtAdjustment
 
 	if adjustedScore >= 4.5 {
 		return "Exceptional"
